datahub/sample-01/requirements.txt
pandas==1.5.3
azure-eventhub==5.11.2
